requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
pkg-resources==0.0.0
pyserial==3.4
pyserial-asyncio==0.4
typing-extensions==3.7.2
urllib3==1.24.3
yarl==1.3.0
